+++ sdks/outpost-typescript/RELEASES.md
@@ -109,3 +109,13 @@ Based on:
 - [typescript v0.10.0] sdks/outpost-typescript
 ### Releases
 - [NPM v0.10.0] https://www.npmjs.com/package/@hookdeck/outpost-sdk/v/0.10.0 - sdks/outpost-typescript
+
+## 2026-04-21 19:27:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.10.1] sdks/outpost-typescript
+### Releases
+- [NPM v0.10.1] https://www.npmjs.com/package/@hookdeck/outpost-sdk/v/0.10.1 - sdks/outpost-typescript